+++ tests/test_kernel.py
@@ -104,7 +104,7 @@ def test_halfspace():                                            # 7. halfspace
     full = kernel.halfspace(**hs['21'])
 
     # Check halfspace = sum of split
-    hs['21']['output'] = 'split'
+    hs['21']['solution'] = 'dsplit'
     direct, reflect, air = kernel.halfspace(**hs['21'])
     assert_allclose(full, direct+reflect+air)
 
@@ -115,15 +115,15 @@ def test_halfspace():                                            # 7. halfspace
     hsbp['21']['aniso'] = hsbp['21']['aniso'][1]
     hsbp['21']['ft'] = 'sin'
     hs_res = bipole(**hsbp['21'])
-    assert_allclose(direct, hs_res)
+    assert_allclose(direct, hs_res, rtol=1e-2)
 
     # Additional checks - Frequency
-    hs['11']['output'] = 'fs'
+    hs['11']['solution'] = 'dfs'
     full = kernel.halfspace(**hs['11'])
 
     # Check halfspace = sum of split
-    hs['11']['output'] = 'fs'
-    direct = kernel.halfspace(**hs['11'])
+    hs['11']['solution'] = 'dsplit'
+    direct, _, _ = kernel.halfspace(**hs['11'])
     assert_allclose(full, direct)
 
     # Check fullspace = bipole-solution
@@ -132,4 +132,4 @@ def test_halfspace():                                            # 7. halfspace
     hsbp['11']['res'] = hsbp['11']['res'][1]
     hsbp['11']['aniso'] = hsbp['11']['aniso'][1]
     hs_res = bipole(**hsbp['11'])
-    assert_allclose(direct, hs_res)
+    assert_allclose(direct, hs_res, atol=1e-2)